requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
tf-keras  # Required for DeepFace with TensorFlow 2.20+

# Face Recognition - Using OpenCV (fast and real-time)
opencv-contrib-python>=4.8.0  # Includes face recognition module

# Pose Estimation - Using simplified OpenCV-based approach
# mediapipe>=0.10.0  # Disabled due to compatibility issues with Python 3.13

# Tracking
deep-sort-realtime>=1.3.2
scipy>=1.10.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
